move carriersSelector out of components dir according to new structure
